requirements.txt
streamlit>=1.33.0
requests>=2.31.0
PyYAML>=6.0.1
scikit-learn>=1.3.2
rapidfuzz>=3.6.1
sentence-transformers>=2.2.2
pypdf>=4.2.0
markdown>=3.6
